+++ domain/src/main/java/woowacourse/shopping/repository/CartRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.repository
 
 import woowacourse.shopping.model.Product
